requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==3.14.0
web3==6.10.0
channels==4.0.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-django==4.5.2
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2
channels-redis==4.0.0
daphne==4.0.0